feat(migration): re-backfill mongo subtractions to postgres
Add a new revision that re-runs the idempotent subtraction backfill to
capture subtractions written or modified between the original
ztd04qgecm9a backfill and the dual-write rollout. The original revision
is already recorded as applied and cannot re-run under its own id.

Extract the shared copy logic into virtool/subtractions/migration.py so
both revisions delegate to a single implementation instead of
duplicating it.
